backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.26",
    "langchain-community>=0.3.26",
    "langchain-openai>=0.3.25",
    "langgraph>=0.4.8",
    "langgraph-cli>=0.3.3",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.8",
]

[dependency-groups]
dev = [
    "mypy>=1.16.1",
]
